Portable/requirements.txt
numpy==1.17.4
plotly==5.5.0
py_cpuinfo==8.0.0
selenium==3.141.0
